Project organisation and some minor changes.
