+++ main.py
@@ -1,26 +1,7 @@
 import discord
-import os
-my_secret = os.environ['TOKEN']
-import requests
-import json
-import random
+from discord_bot import tools
 
 client = discord.Client()
-
-sad_words = ["sad", "depressed", "unhappy", "angry", "miserable", "depressing"]
-
-starter_encouragements = [
-  "Cheer Up",
-  "Hang in there",
-  "You are a great person/bot"
-]
-
-def get_quote():
-  response = requests.get("https://zenquotes.io/api/random")
-
-  json_data = json.loads(response.text)
-  quote = json_data[0]['q'] + " -" + json_data[0]['a']
-  return quote
 
 @client.event
 async def on_ready():
@@ -33,13 +14,13 @@ async def on_message(message):
     return
   
   if message.content.startswith("$inspire"):
-    quote = get_quote()
+    quote = tools.get_quote()
     await message.channel.send(quote)
 
   msg = message.content
   
-  if any(word in msg for word in sad_words):
-    await message.channel.send(random.choice(starter_encouragements))
+  if any(word in msg for word in tools.sad_words):
+    await message.channel.send(tools.get_random_starter_encouragement())
 
   if message.content.startswith("$rishi"):
     await message.channel.send("Rishi you are the best!")
@@ -50,7 +31,7 @@ async def on_message(message):
   if message.content.startswith("$PoSeldOn"):
     await message.channel.send("PoSeldOn you suck")
 
-client.run(my_secret)
 
-  
 
+if __name__ == '__main__':
+  client.run(tools.my_secret)
